Fix name of statements pgpool defines
Postgres limits statement names to around 32 characters. Using
the query itself for the statement name therefore can lead to
conflicts for queries that have the same first 32 characters.

This solution uses erlang:phash2/1 which produces a unique
integer for the given query and uses that integer as part
of the statement name.